+++ bin/lcov-server.js
@@ -2,7 +2,6 @@
 
 require('babel-polyfill');
 
-const program = require('commander');
 const updateNotifier = require('update-notifier');
 
 const cli = require('../lib/cli');
@@ -11,21 +10,73 @@ const pkg = require('../package.json');
 
 updateNotifier({pkg}).notify();
 
-program
-  .version(pkg.version)
-  .option('-u, --upload [server]', 'Set the url to upload lcov data too', 'http://localhost:8080')
-  .option('-s, --serve', 'Pass this option to startup a lcov-server instance')
-  .option('-d, --db [db]', 'Set the db connection', 'mongodb://localhost:32768/lcov-server')
-  .option('-p, --parser <parser>', 'Set the parser value [lcov, cobertura, golang, jacoco], defaults to lcov', 'lcov')
-  .option('-bp, --basePath <path>', 'The path that defines the base directory where the files that were covered will be located')
-  .parse(process.argv);
+const args = process.argv.slice(2);
+
+let program = {};
+
+args.forEach((a, i) => {
+  switch(a) {
+  case '-v':
+  case '--version':
+      console.log(`v${require('../package.json').version}`); // eslint-disable-line
+      process.exit(0);
+    break;
+  case 'help':
+  case '-h':
+  case '--help':
+    console.log(``+ // eslint-disable-line
+  `
+    Usage: lcov-server [options]
+
+    Commands:
+    
+      upload, --upload, -u [server ] Set the url to upload lcov data too (default: http://localhost:8080)
+      serve, -s, --serve             Pass this option to startup a lcov-server instance
+      version, -v, --version           output the version number
+      help, -h, --help              output usage information
+
+    Options:
+
+      db, -d, --db [db]           Set the db connection (default: mongodb://localhost:32768/lcov-server)
+      parser, -p, --parser <parser>   Set the parser value [lcov, cobertura, golang, jacoco], defaults to lcov (default: lcov)
+      basePath, -bp, --basePath <path>  The path that defines the base directory where the files that were covered will be located
+  `);
+    process.exit(0);
+    break;
+    case '-db':
+    case '--db':
+    case 'db':
+      program.db = args[i + 1];
+    break;
+    case '-u':
+    case '--upload':
+    case 'upload':
+      program.upload = args[i + 1];
+    break;
+    case '-s':
+    case '--serve':
+    case 'serve':
+      program.serve = true;
+    break;
+    case '-p':
+    case '--parser':
+    case 'parser':
+      program.parser = args[i + 1];
+      if(['lcov', 'cobertura', 'golang', 'jacoco'].indexOf(program.parser) === -1) {
+        console.error(`parser ${program.parser} not supported`); // eslint-disable-line
+        process.exit(1);
+      }
+    break;
+    case '-bp':
+    case '--basePath':
+    case 'basePath':
+      program.basePath = args[i + 1];
+    break;
+  }
+});
 
 const { parser, upload, serve, db, basePath } = program;
 
-if(parser && ['lcov', 'cobertura', 'golang', 'jacoco'].indexOf(parser) === -1) {
-  console.error(`parser ${parser} not supported`); // eslint-disable-line
-  process.exit(1);
-}
 
 if(serve) {
   process.env.MONGO_URL = process.env.MONGO_URL || db;